+++ src/app/(route)/(withLayout)/layout.tsx
@@ -1,5 +1,6 @@
 import Footer from '@/app/_components/layout/Footer'
 import Header from '@/app/_components/layout/Header'
+import SearchHeader from '@/app/_components/layout/search/SearchHeader'
 import { CookiesProvider } from 'next-client-cookies/server'
 import React, { ReactNode } from 'react'
 import HomeHeader from './_components/HomeHeader'
@@ -8,6 +9,7 @@ export default function WithLayout({ children }: { children: ReactNode }) {
   return (
     <main className="my-0 flex flex-col">
       <CookiesProvider>
+        <SearchHeader />
         <Header />
         <HomeHeader />
         <div className="mb-[18px]">{children}</div>
